% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGameData.R
\name{genLegBetMoveMaxEVDataNTimes}
\alias{genLegBetMoveMaxEVDataNTimes}
\title{Title}
\usage{
genLegBetMoveMaxEVDataNTimes(n)
}
\arguments{
\item{n}{number of games to play}
}
\value{

}
\description{
Title
}
\examples{
x <- genRandomDataNTimes(10, "random", getRandomChoiceAgent, "random", getRandomChoiceAgent)
}
