% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGameData.R
\name{genRandomDataNTimes}
\alias{genRandomDataNTimes}
\title{Title}
\usage{
genRandomDataNTimes(n, agent1Name, agent1FUN, agent2Name, agent2FUN)
}
\arguments{
\item{n}{number of games to play}
}
\value{

}
\description{
Title
}
\examples{
x <- genRandomDataNTimes(10, "random", getRandomChoiceAgent, "random", getRandomChoiceAgent)
}
